test(app): express Codex restore specs as user journeys
